chore: refactor according to code review comments
